+++ client/tests/testdemoboardapi.js
@@ -1057,4 +1057,32 @@ describe("demo board api", () => {
       });
     });
   });
+
+  it("should delete tasks", (done) => {
+    const view = {setState: expect.createSpy()};
+    new BoardAPI(view, 'test2', (api) => {
+      const model = api.model;
+      const task_id = model.subtasks('Backlog')[0].subtasks('ready')[0].id;
+      api.delete(task_id, (api, data) => {
+        expect(data.tasks.removals).toEqual([task_id]);
+        expect(model.subtasks('Backlog')[0].subtasks('ready')).toEqual([]);
+        expect(model.all_tasks.length).toBe(1);
+        done();
+      });
+    });
+  });
+
+  it("should delete features", (done) => {
+    const view = {setState: expect.createSpy()};
+    new BoardAPI(view, 'test2', (api) => {
+      const model = api.model;
+      const feature_id = model.subtasks('Backlog')[0].id;
+      const task_id = model.subtasks('Backlog')[0].subtasks('ready')[0].id;
+      api.delete(feature_id, (api, data) => {
+        expect(data.tasks.removals).toEqual([feature_id, task_id]);
+        expect(model.all_tasks.length).toBe(0);
+        done();
+      });
+    });
+  });
 });