+++ server/twotieredkanban/tests/testboard.py
@@ -663,6 +663,28 @@ class BoardTests(unittest.TestCase):
                               state="Backlog", archived=True,
                               ), vars.p1.history[-2])
 
+    def test_remove_task(self):
+        self.board.new_project("Feature", 0)
+        [fid] = [t.id for t in self.updates()['tasks']['adds']]
+        self.board.new_task(fid, "t1", 0)
+        self.board.new_task(fid, "t2", 0)
+        t1id, t2id = [t.id for t in self.updates()['tasks']['adds']]
+        self.board.remove(t1id)
+        self.assertEqual([t1id], self.updates()['tasks']['removals'])
+        self.assertEqual(len(self.board.tasks), 2)
+
+    def test_remove_feature(self):
+        self.board.new_project("Feature", 0)
+        [fid] = [t.id for t in self.updates()['tasks']['adds']]
+        self.board.new_task(fid, "t1", 0)
+        self.board.new_task(fid, "t2", 0)
+        t1id, t2id = [t.id for t in self.updates()['tasks']['adds']]
+        self.board.remove(fid)
+        self.assertEqual(
+            set([fid, t1id, t2id]),
+            set(self.updates()['tasks']['removals']))
+        self.assertEqual(len(self.board.tasks), 0)
+
 
 sample_description = """
 <h1>Heading large</h1>